msm8996-common: Init ConfigPanel as soon as possible
It was recently noticed that waiting for BOOT_COMPLETED intent action
is not enough to make sure that stuff is properly initialized
before the user opens additional buttons settings. Address such issue
by registering to INITIALIZE_LINEAGE_HARDWARE intent action as well,
but keep listening to BOOT_COMPLETED action just in case the SDK
is not present or not brought up.

Signed-off-by: xNombre <kartapolska@gmail.com>